refactor: UserProfilePage 코드 개선 (#9)
