+++ ensolution-front/src/features/auth/ui/index.ts
@@ -1,3 +1,4 @@
 export * from './LoginForm';
+export * from './ProfileForm';
 export * from './SignupHint';
 export * from './SocialLoginButton';
